server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
pyJWT==2.8.0
flask-session==0.5.0